[BEAM-2211] Move PathValidator into GCP-Core
For now, this does not need to be a Beam concept
